revamp getHistoricalTransactions
